+++ graphbrain/__init__.py
@@ -1,5 +1,12 @@
+import os
+
 from graphbrain.hyperedge import hedge
-import graphbrain.memory.leveldb
+import graphbrain.memory.sqlite
+try:
+    import graphbrain.memory.leveldb
+    LEVELDB = True
+except ImportError:
+    LEVELDB = False
 
 
 def hgraph(locator_string):
@@ -9,4 +16,12 @@ def hgraph(locator_string):
     Currently, the only type of hypergraph is based on LevelDB storage,
     and the location_string is the path to the LevelDB folder.
     """
-    return graphbrain.memory.leveldb.LevelDB(locator_string)
+    if os.path.isfile(locator_string): 
+        return graphbrain.memory.sqlite.SQLite(locator_string)
+    elif os.path.isdir(locator_string):
+        if LEVELDB:
+            return graphbrain.memory.leveldb.LevelDB(locator_string)
+        else:
+            print('LevelDB not supported. Custom build required.')
+    else:
+        print('Hypergraph database could not be found.')