+++ graphbrain/memory/sqlite.pyx
@@ -1,6 +1,6 @@
 import json
 
-import plyvel
+import sqlite3
 
 from graphbrain.hyperedge import edges2str
 from graphbrain.hyperedge import hedge
@@ -14,63 +14,66 @@ from graphbrain.memory.permutations import str_plus_1
 
 
 def _edge2key(edge):
-    return (''.join(('v', edge.to_str()))).encode('utf-8')
+    return edge.to_str()
 
 
 def _encode_attributes(attributes):
     return json.dumps(attributes,
                       ensure_ascii=False,
                       check_circular=False,
-                      separators=(',', ':')).encode('utf-8')
+                      separators=(',', ':'))
 
 
 def _decode_attributes(value):
-    return json.loads(value.decode('utf-8'))
+    return json.loads(value)
 
 
-class LevelDB(Hypergraph):
-    """Implements LevelDB hypergraph storage."""
+class SQLite(Hypergraph):
+    """Implements SQLite hypergraph storage."""
 
     def __init__(self, locator_string):
         self.locator_string = locator_string
-        self.db = plyvel.DB(self.locator_string, create_if_missing=True)
+        self.conn = sqlite3.connect(self.locator_string, isolation_level=None)
+        self.conn.execute(
+            'CREATE TABLE IF NOT EXISTS "v" '
+            '(key TEXT PRIMARY KEY, value TEXT)')
+        self.conn.execute(
+            'CREATE TABLE IF NOT EXISTS "p" (key TEXT PRIMARY KEY)')
+        self.conn.execute(
+            'CREATE TABLE IF NOT EXISTS "c" (key TEXT PRIMARY KEY, value INT)')
 
     # ============================================
     # Implementation of abstract interface methods
     # ============================================
 
     def close(self):
-        self.db.close()
+        if self.conn:
+            self.conn.close()
+            self.conn = None
 
     def name(self):
         return self.locator_string
 
     def destroy(self):
-        self.db.close()
-        plyvel.destroy_db(self.locator_string)
-        self.db = plyvel.DB(self.locator_string, create_if_missing=True)
+        cur = self.conn.cursor()
+        cur.execute('DELETE FROM v')
+        cur.execute('DELETE FROM p')
+        cur.execute('DELETE FROM c')
 
     def all(self):
-        start_str = 'v'
-        end_str = str_plus_1(start_str)
-        start_key = start_str.encode('utf-8')
-        end_key = end_str.encode('utf-8')
-
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            edge = hedge(key.decode('utf-8')[1:])
+        cur = self.conn.cursor()
+        for key, _ in cur.execute('SELECT * FROM v'):
+            edge = hedge(key)
             if edge is not None:
                 yield edge
 
     def all_attributes(self):
-        start_str = 'v'
-        end_str = str_plus_1(start_str)
-        start_key = start_str.encode('utf-8')
-        end_key = end_str.encode('utf-8')
-
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            edge = hedge(key.decode('utf-8')[1:])
-            attributes = _decode_attributes(value)
-            yield (edge, attributes)
+        cur = self.conn.cursor()
+        for key, value in cur.execute('SELECT * FROM v'):
+            edge = hedge(key)
+            if edge is not None:
+                attributes = _decode_attributes(value)
+                yield (edge, attributes)
 
     def atom_count(self):
         return self._read_counter('atom_count')
@@ -162,17 +165,18 @@ class LevelDB(Hypergraph):
                     positions.append(i)
             start_str = edges2str(nodes)
             end_str = str_plus_1(start_str)
-            start_key = (''.join(('p', start_str))).encode('utf-8')
-            end_key = (''.join(('p', end_str))).encode('utf-8')
 
-            for key, value in self.db.iterator(start=start_key, stop=end_key):
-                perm_str = key.decode('utf-8')
-
-                tokens = split_edge_str(perm_str[1:])
+            cur = self.conn.cursor()
+            for row in cur.execute(
+                    'SELECT * FROM p WHERE key >= ? AND key < ?',
+                    (start_str, end_str)):
+                key = row[0]
+                tokens = split_edge_str(key)
                 nper = int(tokens[-1])
+                print(nper)
 
                 if nper == first_permutation(len(tokens) - 1, positions):
-                    yield perm2edge(perm_str[1:])
+                    yield perm2edge(key)
 
     def _match(self, pattern, strict=True, curvars={}):
         for edge in self._match_structure(pattern, strict):
@@ -184,18 +188,19 @@ class LevelDB(Hypergraph):
         center_str = center.to_str()
         start_str = ''.join((center_str, ' '))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('p', start_str))).encode('utf-8')
-        end_key = (''.join(('p', end_str))).encode('utf-8')
 
         count = 0
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
+        cur = self.conn.cursor()
+        for row in cur.execute(
+                'SELECT * FROM p WHERE key >= ? AND key < ?',
+                (start_str, end_str)):
+            key = row[0]
             if limit and count >= limit:
                 break
-            perm_str = key.decode('utf-8')
-            edge = perm2edge(perm_str[1:])
+            edge = perm2edge(key)
             if edge:
                 position = edge.index(center)
-                nper = int(split_edge_str(perm_str[1:])[-1])
+                nper = int(split_edge_str(key)[-1])
                 if nper == first_permutation(len(edge), (position,)):
                     count += 1
                     yield(edge)
@@ -203,11 +208,12 @@ class LevelDB(Hypergraph):
     def _atoms_with_root(self, root):
         start_str = ''.join((root, '/'))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('v', start_str))).encode('utf-8')
-        end_key = (''.join(('v', end_str))).encode('utf-8')
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            symb = hedge(key.decode('utf-8')[1:])
+        cur = self.conn.cursor()
+        for key, value in cur.execute(
+                'SELECT * FROM v WHERE key >= ? AND key < ?',
+                (start_str, end_str)):
+            symb = hedge(key)
             yield(symb)
 
     def _edges_with_edges(self, edges, root):
@@ -215,17 +221,18 @@ class LevelDB(Hypergraph):
         if root:
             start_str = ''.join((start_str, ' ', root, '/'))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('p', start_str))).encode('utf-8')
-        end_key = (''.join(('p', end_str))).encode('utf-8')
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            perm_str = key.decode('utf-8')
-            edge = perm2edge(perm_str[1:])
+        cur = self.conn.cursor()
+        for row in cur.execute(
+                'SELECT * FROM p WHERE key >= ? AND key < ?',
+                (start_str, end_str)):
+            key = row[0]
+            edge = perm2edge(key)
             if edge:
                 if root is None:
                     if all([item in edge for item in edges]):
                         positions = [edge.index(item) for item in edges]
-                        nper = int(split_edge_str(perm_str[1:])[-1])
+                        nper = int(split_edge_str(key)[-1])
                         if nper == first_permutation(len(edge), positions):
                             yield(edge)
                 else:
@@ -269,12 +276,16 @@ class LevelDB(Hypergraph):
     def _add_key(self, key, attributes):
         """Adds the given edge, given its key."""
         value = _encode_attributes(attributes)
-        self.db.put(key, value)
+        cur = self.conn.cursor()
+        cur.execute(
+            'INSERT INTO v (key, value) VALUES(?, ?) '
+            'ON CONFLICT(key) DO UPDATE SET value=excluded.value',
+            (key, value))
 
     def _write_edge_permutation(self, perm):
         """Writes a given permutation."""
-        perm_key = (''.join(('p', perm))).encode('utf-8')
-        self.db.put(perm_key, b'x')
+        cur = self.conn.cursor()
+        cur.execute('INSERT INTO p (key) VALUES(?)', (perm,))
 
     def _write_edge_permutations(self, edge):
         """Writes all permutations of the edge."""
@@ -282,8 +293,8 @@ class LevelDB(Hypergraph):
 
     def _remove_edge_permutation(self, perm):
         """Removes a given permutation."""
-        perm_key = (''.join(('p', perm))).encode('utf-8')
-        self.db.delete(perm_key)
+        cur = self.conn.cursor()
+        cur.execute('DELETE FROM p WHERE key = ?', (perm,))
 
     def _remove_edge_permutations(self, edge):
         """Removes all permutations of the edge."""
@@ -291,11 +302,15 @@ class LevelDB(Hypergraph):
 
     def _remove_key(self, key):
         """Removes an edge, given its key."""
-        self.db.delete(key)
+        cur = self.conn.cursor()
+        cur.execute('DELETE FROM v WHERE key = ?', (key,))
 
     def _exists_key(self, key):
         """Checks if the given edge exists."""
-        return self.db.get(key) is not None
+        cur = self.conn.cursor()
+        for key, _ in cur.execute('SELECT * FROM v WHERE key = ?', (key, )):
+            return True
+        return False
 
     def _set_attribute_key(self, key, attribute, value):
         """Sets the value of an attribute by key."""
@@ -336,8 +351,10 @@ class LevelDB(Hypergraph):
             return False
 
     def _attribute_key(self, key):
-        value = self.db.get(key)
-        return _decode_attributes(value)
+        cur = self.conn.cursor()
+        for key, value in cur.execute('SELECT * FROM v WHERE key = ?', (key,)):
+            return _decode_attributes(value)
+        return None
 
     def _get_str_attribute_key(self, key, attribute, or_else=None):
         if self._exists_key(key):
@@ -369,29 +386,27 @@ class LevelDB(Hypergraph):
         else:
             return or_else
 
-    def __read_counter_key(self, counter_key):
-        """Reads a counter by key."""
-        value = self.db.get(counter_key)
-        if value is None:
-            return 0
-        else:
-            return int(value.decode('utf-8'))
-
     def _read_counter(self, counter):
         """Reads a counter by name."""
-        return self.__read_counter_key(counter.encode('utf-8'))
+        cur = self.conn.cursor()
+        for key, value in cur.execute(
+                'SELECT * FROM c WHERE key = ?', (counter,)):
+            return value
+        return 0
 
     def _inc_counter(self, counter, by=1):
         """Increments a counter."""
-        counter_key = counter.encode('utf-8')
-        value = self.__read_counter_key(counter_key)
-        self.db.put(counter_key, str(value + by).encode('utf-8'))
+        cur = self.conn.cursor()
+        cur.execute(
+            'INSERT INTO c (key, value) VALUES(?, ?) '
+            'ON CONFLICT(key) DO UPDATE SET value = value + excluded.value',
+            (counter, by))
 
     def _dec_counter(self, counter, by=1):
         """Decrements a counter."""
-        counter_key = counter.encode('utf-8')
-        value = self.__read_counter_key(counter_key)
-        self.db.put(counter_key, str(value - by).encode('utf-8'))
+        cur = self.conn.cursor()
+        cur.execute(
+            'UPDATE c SET value = value - ? WHERE key = ?', (by, counter))
 
     def _inc_degrees(self, edge, depth=0):
         if depth > 0: